+++ signing_service/signing_service/tests/test_authenticate.py
@@ -15,6 +15,7 @@ from middleman_protocol.stream import unescape_stream
 
 from signing_service.constants import SIGNING_SERVICE_DEFAULT_RECONNECT_ATTEMPTS
 from signing_service.signing_service import SigningService
+from signing_service.utils import ConsoleNotifier
 from .utils import SigningServiceIntegrationTestCase
 
 
@@ -125,6 +126,7 @@ class TestSigningServiceAuthenticate(SigningServiceIntegrationTestCase):
                         SIGNING_SERVICE_PRIVATE_KEY,
                         TEST_ETHEREUM_PRIVATE_KEY,
                         SIGNING_SERVICE_DEFAULT_RECONNECT_ATTEMPTS,
+                        ConsoleNotifier(),
                     )
 
                     # For test purposes we reverse roles, so signing service works as server.