+++ signing_service/signing_service/tests/test_handle_connection.py
@@ -26,6 +26,7 @@ from signing_service.constants import SIGNING_SERVICE_DEFAULT_RECONNECT_ATTEMPTS
 from signing_service.constants import WARNING_DAILY_THRESHOLD
 from signing_service.exceptions import SigningServiceValidationError
 from signing_service.signing_service import SigningService
+from signing_service.utils import ConsoleNotifier
 from .utils import SigningServiceIntegrationTestCase
 
 
@@ -297,6 +298,7 @@ class TestSigningServiceHandleConnection(SigningServiceIntegrationTestCase):
                         SIGNING_SERVICE_PRIVATE_KEY,
                         TEST_ETHEREUM_PRIVATE_KEY,
                         SIGNING_SERVICE_DEFAULT_RECONNECT_ATTEMPTS,
+                        ConsoleNotifier(),
                     )
 
                     # For test purposes we reverse roles, so signing service works as server.